pyproject.toml
[project]
name = "moobie-time"
version = "0.1.0"
description = "Discord bot to manage weekly movie night suggestions and watch list"
readme = "README.md"
requires-python = ">=3.14"
dependencies = ["discord.py", "pydantic", "sqlalchemy", "tvdb_v4_official"]
